requirements.txt
python-docx
openpyxl
python-pptx
reportlab
Aspose.Email-for-Python-via-NET
